Back link should accept custom html attributes
Inital I thought of using native html but this proved to be troublesome
when it came to introducing custom html attributes. I did think about just
a hash but again this was a bit fiddly.

The only difference now between this component and govuk-frontend
design system version is that their back link accepts custom html
to be passed as link text. I'm not sure I agree with this because
theres risk that user could try pass it a link which would be invalid
to have an link inside a link. I think for the moment we can afford
not to support it until someone requests it.