Add the tab component that utilises the Slot API
The ViewComponent Slot API is the experimental replacement for
ContentArea. It offers *far* more power and lets us repeat (via
`collection`) calls to slots, so we can define an unlimited number of
tabs with a clean DSL-like syntax.

The main bonus here, and the cause of the delay on this work, is that
now testing block functionality is fully supported by `render_inline`
taking a block.