Allow slots to be wrapped by methods on component
This makes interaction with the components library even more DSL-like
and reduces the amount of code in views considerably. Now, instead of
using the `render` syntax and calling the components by class name, a
handy helper method is provided.

Additionally, the `wrap_slot` macro adds a method with the name of the
slot to the component, so like in this example, sections can be added
using `accordion.section` rather than `accordion.slot(:section)`.

govuk_accordion(id: 'def234') do |accordion|
  accordion.section(title: 'Section 1') do
    tag.p(class: 'govuk-body') do
      "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
    end
  end
end