Rename methods added to simplify adding slots
The dynamic addition of methods to allow for 'DSL-style' slots causes a
problem when the component is a collection. The problem is that when
`collection: true` on the outer slot, a singular method is added to the
component instance (so `accordion.item` would be defined already and
then we'd blindly overwrite it).

As the upstream API hasn't settled yet, the temporary 'fix' is to prefix
the method with `add_`, which in most cases will make sense.

The alternative syntax of using `component.slot(:item, args)` is
unaffected.

Refs #63