Allow accordion heading to be set with custom HTML
This change adds a heading_html slot to the accordion section that takes
a block of HTML. If both heading_text and heading_html are supplied the
HTML will take precedence. If neither are supplied an error will be
raised.

To ease the process of testing the accordion has been split out into a
separate file; I repeatedly ran into problems getting the tests to work
when the accordion and section components shared a single file[0].

A lambda slot was used because ideally we want to specify the heading
level for each section once in the accordion; it wouldn't be ideal to
have to do it once-per-section. Lambda slots are the only style that
support sharing values between the parent and child.

[0] https://github.com/github/view_component/issues/1000