Rename header items to navigation items
In the Nunjucks macros a single array of navigation objects is passed
in. Here, we'd need to call the slot multiple times so #navigation_item
makes more sense than #navigation. The individual components of
the navigation menu in the Design System are referred to as items so I
think this is the best compromise.