Rewrite the cookie banner with slots
This now works in a very similar fashion to the Nunjucks macros. Instead
of passing in actions with arguments it supports the passing in of any
arbitrary HTML, the intention is for developers to pass in an array of
buttons or links generated by `govuk_link_to` or similar. These will
then automatically be wrapped in a `govuk-button-group`.