Begin work on the table component
This is going to be the most complex component here. After some
experimentation it's likely that not all the Nunjucks options will be
supported out-of-the-box, mainly due to the inability for slots to
access their parents.

The entire set of functionality (and more) will be available when
building via slots, however.