Fix breadcrumb text entries when key is symbol
When breadcrumb keys (the text) are symbols and there is no link, the
rendering is broken because of the way ActiveView::Helpers::CaptureHelper
works.

Now we're calling #to_s on the link text to ensure it renders. Also
added a line to the documentation explaining that an empty string can be
used here too.

Refs #249
Fixes #250