Render some examples in iframes
This solves two problems:

Firstly, the JavaScript that triggers the menu behaviour for headers
only targets _the first header_ it encounters. This makes sense as in
normal pages there'll never be more than one. Here it is problematic as
we have several. The header being in a iframe means that it's in a
separate context so the JavaScript runs again.

Secondly, we can use sandbox mode on the iframe to stop JavaScript from
focussing the element on load. This needs to be configured manually
because we _want_ JavaScript in the header example but not in the
notification banner ones