Add actions keyword to summary list component
Version 4.0.0 of the design system changes the behaviour of summary
lists. Instead of adding an empty actions column the
.govuk-summary-list__row--no-actions class is applied to the row and it
will add a column as a pseudoelement via CSS.

As we're building lists in a DSL fashion, our nodes don't know about
their siblings; we can't work out whether any other rows have actions.
We need to _tell_ the summary list how to act and it will apply the
classes automatically.

This work also paves the way for an alternative syntax where we pass in
an array of hashes/objects - this would work how the table works. In
this case (because we can see the objects up front) we _can_ take care
of this automatically.