Switch to Dart Sass
GOV.UK Frontend is dropping support for Ruby Sass soon so we may as well
make the jump to Dart Sass sooner rather than later.

https://github.com/alphagov/govuk-frontend/discussions/2632

Here we're integrating it with Nanoc using a custom filter, if Nanoc
adds native support at some point it'd make sense to switch to it as
this isn't the cleanest thing ever and is totally inflexible, but the
approach should be fine for here and the form builder.