Add a section break component
In the Typography section of the design system, there is a part on the [Section
break element](https://design-system.service.gov.uk/styles/typography/#section-break).

This adds a component (and respective helper method) for rendering a section
break element, as suggested in #321.

The visibility and size of the component is configurable.