Handle nil classes more gracefully
When the classes param is nil the library fails while attempting to
append conditional classes to it. The error message doesn't make it
obvious what the problem is.

Here we are addressing it by emitting a warning when nil classes are
detected and overwriting the nil with an empty array.

Fixes #334