Make pagy a dependency
When Zeitwerk attempts to eager_load dependencies it fails unless Pagy
is present. This wasn't apparent in regular tests as eager load is
disabled but was uncovered in a CI run because of the config

  config.eager_load = ENV["CI"].present?

In the short term it's easier to add Pagy as a dependency than try to
work around Zeitwerk.