Add support for array of meta item hashes
The original implementation of meta items wasn't feature complete with
the GOV.UK original in that it took a hash of `{ text: href }` style
meta items.

The GOV.UK frontend version instead takes an array of hashes
`[{ href:, text:, attr: }]` It's more powerful because it allows
custom attributes to be set for each link.

Here we're adding support for original format in addition to the
condensed one.