Build basic card functionality for summary lists
The summary list card is an upcoming addition to the summary list
component that renders it inside a card with a set of list-wide actions
at the top alongside a header.

This is a unusual component because usually the component 'root' is at
the outermost level (e.g., `govuk-header`) and everything within uses
the namespace root as a prefix (BEM style, e.g., `govuk-header__logo`)

Here though, the `govuk-summary-list__card` is outside the
`govuk-summary-list`, so there's a little awkwardness in how it's
called:

* We can either take the _conventional_ ViewComponent route and do:
  = govuk_summary_list_card(title: "My card") do |card|
    - card.action(...)
    - card.summary_list(govuk_summary_list(...))

* Or we can take the GOV.UK Design System route and do:
  = govuk_summary_list(header: { title: "My card" }) do |sl|
    - sl.row(...)
    - sl.row(...)

Examples of both have been added to the guide.