Require visually_hidden_text to be set on summary list actions
Visually hidden text is nearly always required on the action links for summary list items, as otherwise they will often all just contain "Change", which is unhelpful for screen reader users who may hear links read out out-of-context of the rest of the page.

Unfortunately, it’s relatively common to forget to set this, and for this not to be picked up when reviewing code or the service.

This change aims to make this accessibility issue less common by requiring `visually_hidden_text` to be set on the action items of summary list rows. If it is not set, ruby will throw an error.

In the rare occasions where visually-hidden text is not required (say for instance there's only a single action link on the page, or all of the links have different link text), `visually_hidden_text` can be explicitly set to `nil`.