Add scopes attr and default values to table header
Table header cells `th` can appear in two places:

* Usually they're only present in the `thead` as column headers so they
  have the attribute `scope='col'`.
* Sometimes they are also used in the first column to mark the row
  header, then they have the attribute `scope='row'`.

This commit adds a 'scope' parameter to the CellComponent. It's a
required parameter but it's automatically set in the `#cell_scope`
method at the row level if it's missing, so unless you're using the
CellComponent standalone you should never have to set it.

To determine the kind of scope we also need to know whether the
grandparent of the cell is the `thead` or `tbody`, so `parent` has been
added as a parameter to the RowComponent. Again, it will need to be
manually set if the component is used standalone but when used within a
table it's automatically assigned.