Reorganise and move logic to the cell
Trying to make the row responsible for the cell's default scope was
problematic because should we want to override it at the cell level we
don't know whether it was set manually or defaulted.

I originally tried this by passing another param but it's cleaner to
just move the logic to the cell and pass the 'parent' down.

Not sure parent is the best name for it, perhaps 'region'. At the second
level it's actually a grandparent.