Set table header default default by location
This setting allows us to remove the header param from several places as
we can tidily infer whether a cell should be a `th` or a `td` by whether
its parent is a `thead` or a `tbody` while still respecting the
first_cell_is_header setting and any manual overrides.

The default is applied at the cell level whenever `header: nil`, when
it's truthy or falsy the override will take precedence.