Add support for table colgroups and cols
The `<colgroup>` element allows us to group related columns `<col>`
together in a table. They are useful when building complex tables that
have more than one row of headers.

This change adds ColGroupComponent and ColGroupComponent::ColComponent
to tables, allowing us to specify both in a fashion similar to rows and
cells.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/colgroup