Remove new tab functionality
Despite this being a good move the implementation breaks because of
the close ties to Rails' parameter names. We have to do a lot of
juggling of params to make them flexible enough to accept either a name
and href, a href and block, a name and Rails action/controller via
options etc.

Adding a named param `new_tab` results in the unenclosed
options/extra_options hash being interpreted as a named param, meaning
users of this library would need to go through their app and make
changes like this:

Before:

= link_to("some page", "/some/path", class: "thing")

After:

= link_to("some page", "/some/path", { class: "thing" })

Otherwise, class is treated as a named arg rather than being the options hash.