Move visually hidden space inside element
The final decision on how to properly handle the space between the
visible and hidden text hasn't been made yet but the general consensus
upstream is that it's better inside the visually hidden span than before
it.

This might be revisited again when a decision has been made and the
solution presented in #403 can be used instead.