Add fake Rails app so ViewComponent works properly
A recent change to ViewComponent[0] makes it call
Rails.application.routes.url_helpers which, until now, didn't exist.

We can initialize a fake Rails application so it's at least present,
despite being totally empty, which will allow the guide build to
function. We're very much in the minority in terms of how we're building
this guide Rails-free, so it's on us to make it work.

[0] https://github.com/ViewComponent/view_component/pull/1765