Set the item id numbers sequentially
This change brings the behaviour of the task list more in line with the
Nunjucks component. Now when there's no id_prefix param the ids of the
statuses and tags are just numbered sequentially. If there's going to
be more than one task list on a page they can have their prefixes set
seprately so there's no ID clashes on the final rendered page.

Co-authored-by: Frankie Roberto <frankie@frankieroberto.com>