Fix block syntax for buttons
If you use the block syntax, then the href is actually confusingly the variable named `name` as it is the first argument, rather than the one named `href`.

Unfortunately the tests didn't pick this up as the mock for `url_for` always returned the right variable, even if passed `nil`. Swapping this out for a minimal implementation of `url_for` avoids these false positives.