Switch action_suffix to visually_hidden_action_suffix
Having deliberated a bit over naming I think this is best. Originally it
was `card_title` but I think it's misleading because it doesn't actually
set a card's title, and `card_title: "xyz"` looks quite similar to
`card: { title: "xyz" }`, but they do totally different things.

This name is an improvement over `action_suffix` because it tells the
dev that it's visually hidden and comes after the action. It's inkeeping
with other visually hidden params like `visually_hidden_text` too.