Add new tab text to links that open in new tab
The GOV.UK Design System says this:

> If you need a link to open in a new tab - for example, to stop the
> user losing information they’ve entered into a form - then include the
> words ‘opens in new tab’

However, in the examples the new tab text is surrounded by parentheses.

To reduce repetition here we are changing the behaviour of links that
open in a new tab so that the text is automatically inserted.

The text defaults to `(link opens in new tab)` but this can be
overridden using the `default_link_new_tab_text` configuration option.

The behaviour can be suppressed entirely by setting this value to
`nil`.

On a link-by-link basis, the new tab text can be overridden by passing a
string instead of true to 'new tab', so calling:

<%= govuk_link_to("hello", "/some/page", new_tab: "(opens in new window)") %>

will generate a link that has the overridden text as a suffix. If we
want to visually hide this text it could be called with:

<%= govuk_link_to("hello", "/some/page", new_tab: govuk_visually_hidden("(opens in new window)")) %>