Improve the argument assignment in links
The underlying Rails helpers have two 'modes', on where the link text
and href are passed in as arguments and one where the href is passed in
as an argument and the link text is provided via a block.

This means we have to juggle the two arguments based on the presence of
a block. Until now I left the names as they're defined in the method
signature but that leads to confusing as we were holding the href in a
variable called `link_name`.

Now, the two temporary arguments are called 'a' and 'b', and each link
method's first step is to assign name and href (or name and
email_address for govuk_mail_to) appropriately.