Updates to Tabs documentation
When implementing this I initially didn't spot that the the tab slots would accept a block. Feel like this is probably the default use case (as tabs normally contain a fair amount of content) so worth highlighting this a bit more.

Also added some documentation on the `title` argument, and renamed the local variable from `c` to `tabs`.