Build copyright when rendering instead of at init
ViewComponent raises a ControllerCalledBeforeRenderError when rendering
previews because we're calling #link_to at the point the component is
initialized, which is problematic because the view context only exists
once the ViewComponent is passed to the Rails render pipeline.

Setting copyright_text and copyright_url during initialization but
combining them into a link with link_to when rendering the component
should alleviate this.

Fixes #512