Allow tabs to have set ids
The tab id attribute is currently based on the label given the tab. In
some cases, however, this label may be dynamic, e.g. if the label
includes the number of records included in that tab. Having an id that
changes this way can get in the way of testing or potentially other
functionality that may depend on the id being static.