Make Pagy use the legacy nav bar
Pagy has deprecated the size array format and replaced it with an
integer. This is more sensible but doesn't fit in with the GOV.UK Design
System's guidance that says:

> For smaller screens, show page numbers for:
>
> * the current page
> * previous and next pages
> * first and last pages
>
> For larger screens, show page numbers for:
>
> * the current page
> * at least one page immediately before and after the current page
> * first and last pages

The numeric size doesn't provide quite enough control over what's
displayed before/after the current page.

As a result we have to enable the Size Extra[0] functionality, which
allows us to use the old implementation.

Also, Pagy has replaced the items param with limit[1], so the specs have
been updated accordingly.

[0] https://ddnexus.github.io/pagy/docs/extras/size/
[1] https://ddnexus.github.io/pagy/changelog/#version-900