Allow branding to be configured at component level
There are some components in the GOV.UK Design System that don't appear
in the NHS design system. Currently teams who want to use these
components need to manually copy and adjust chunks of SCSS, replacing
`govuk` for `nhsuk` - a laborious process.

We can make this a little easier by allowing the brand to be configured
on a component-by-component basis, reducing the amount of custom styles
needed.

The reason this approach was taken and not just passing in a `brand`
override to each component is that there's no easy way of passing
arguments through to the slots without converting them all to
lambdas[0].

This will likely be a really niche feature.

[0] https://viewcomponent.org/guide/slots.html#lambda-slots