Allow setting custom licence HTML in footer
There is already an attribute to set custom licence text, however
this does not allow including links or a logo, meaning it's not
possible to internatialise the footer without removing these items.

Allow passing in `licence_meta_html` as a block to be included instead
of the default licence content.