Use correct class names in cell widths
We are using the short/convenience width name as the class for cell widths, so:

```
row.with_cell(text: 'Name', width: 'one-third')
```

Will render out as:

```
<th class="govuk-table__header one-third" scope="col">Name</th>
```

As a result the width is not correctly applied; we need to use the fully
qualified class name from GOV UK frotend.

Update `CellComponent` to convert to the correct class name, for example
`govuk-!-width-one-third`.

Fix an issue with the specs where Nokogiri throws an error due to trying to run
a CSS selectcor with an `!` in it; this needs to be escaped.

Move width/class mappings into an instance method so that `brand` is available.