Fix 'exit this page now' previews
The exit this page now component consists of a div that is set to
'float: right' with a link in it. Floating elements don't contribute to
their container's height so the preview box is squashed.

Using 'contain: layout'[0] isolates the element from the rest of the
page and stops the floating element from overhanging the preview box.

[0] https://developer.mozilla.org/en-US/docs/Web/CSS/contain#layout
[1] https://www.smashingmagazine.com/2019/12/browsers-containment-css-contain-property/