Update links in pagination component
Pagy has had a major update since this page was written and removed all of their old docs in the process, so the links in this document break. The API has changed so we can't point to the new docs, so instead this points to the last version of the v9 documentation on github.